+++ tests/e2e/utils/CHANGELOG.md
@@ -13,7 +13,9 @@
 - Added `itIf()` to conditionally run a test case.
 - Added merchant workflows around plugins: `uploadAndActivatePlugin()`, `activatePlugin()`, `deactivatePlugin()`, `deletePlugin()`
 - Added merchant workflows checking for a database update and performing the update if needed: `runDatabaseUpdate()`
+- Added `deleteAllOrders()` that goes through and deletes all orders
 - Added `deleteAllShippingClasses()` which permanently deletes all shipping classes using the API
+- Added `statuses` optional parameter to `deleteAllRepositoryObjects()` to delete on specific statuses
 
 # 0.1.5